# onidelcloud/onidel-api-docs.yaml
openapi: 3.0.3
info:
  title: Onidel Cloud API - OpenAPI 3.0
  description: |-
    Welcome to the **Onidel Cloud API** documentation. This API allows developers to seamlessly integrate with the Onidel Cloud platform, enabling them to manage cloud resources efficiently and securely. With our API, you can automate processes, retrieve detailed account information, and manage virtual machines (VPS), networking, storage, and more.
  termsOfService: https://docs.onidel.com/policies/terms-of-service
  contact:
    email: support@onidel.com
  version: 1.2.0
servers:
  - url: https://api.cloud.onidel.com
...
paths:
...
  /network/firewalls:
    get:
      tags:
        - Firewall
      summary: List Firewall Groups
      description: Retrieve all firewall groups for a team
      operationId: getFirewalls
      parameters:
        - name: team_id
          in: query
          description: Team ID to filter firewall groups. Default team will be used if not provided.
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FirewallGroups'
        '401':
          description: Unauthorized
      security:
        - api_key: []
    post:
      tags:
        - Firewall
      summary: Create Firewall Group
      description: Create a new firewall group
      operationId: createFirewall
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewFirewallGroup'
      responses:
        '201':
          description: Firewall group created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FirewallGroupDetail'
        '400':
          description: Bad request - missing parameters or limit reached
        '401':
          description: Unauthorized
      security:
        - api_key: []
  /network/firewalls/{firewall_id}:
    get:
      tags:
        - Firewall
      summary: Get Firewall Group by ID
      description: Retrieve a specific firewall group by its UUID
      operationId: getFirewall
      parameters:
        - name: firewall_id
          in: path
          description: Firewall Group UUID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FirewallGroupDetail'
        '401':
          description: Unauthorized
        '404':
          description: Firewall group not found
      security:
        - api_key: []
    put:
      tags:
        - Firewall
      summary: Update Firewall Group
      description: Update a firewall group's description
      operationId: updateFirewall
      parameters:
        - name: firewall_id
          in: path
          description: Firewall Group UUID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFirewallGroup'
      responses:
        '200':
          description: Firewall group updated successfully
        '401':
          description: Unauthorized
        '404':
          description: Firewall group not found
      security:
        - api_key: []
    delete:
      tags:
        - Firewall
      summary: Delete Firewall Group
      description: Delete a firewall group. Cannot delete if VMs are still attached.
      operationId: deleteFirewall
      parameters:
        - name: firewall_id
          in: path
          description: Firewall Group UUID
          required: true
          schema:
            type: string
            format: uuid
        - name: team_id
          in: query
          description: Team ID
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Firewall group deleted successfully
        '400':
          description: Cannot delete - VMs still attached
        '401':
          description: Unauthorized
        '404':
          description: Firewall group not found
      security:
        - api_key: []
...
  /network/ip_lists:
    get:
      tags:
        - IP List
      summary: List IP Lists
      description: Retrieve all IP lists for a team
      operationId: getIPLists
      parameters:
        - name: team_id
          in: query
          description: Team ID. Default team will be used if not provided.
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IPLists'
        '401':
          description: Unauthorized
      security:
        - api_key: []
    post:
      tags:
        - IP List
      summary: Create IP List
      description: Create a new IP list. Limit per team is configurable.
      operationId: createIPList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewIPList'
      responses:
        '201':
          description: IP list created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IPListCreateResponse'
        '400':
          description: Bad request - missing parameters, duplicate name, or limit reached
        '401':
          description: Unauthorized
      security:
        - api_key: []
...
components:
  schemas:
...
    FirewallGroup:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: e4d3c2b1-5678-9abc-def0-fedcba098765
        description:
          type: string
          example: "Production Firewall"
        created:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updated:
          type: string
          format: date-time
          example: "2024-01-20T14:45:00Z"
        instance_count:
          type: integer
          example: 3
          description: "Number of VMs using this firewall group"
        rule_count:
          type: integer
          example: 5
          description: "Number of firewall rules in this group"
      required:
        - id
        - description
        - created
        - updated
        - instance_count
        - rule_count
    FirewallGroups:
      type: object
      properties:
        firewall_groups:
          type: array
          items:
            $ref: '#/components/schemas/FirewallGroup'
    FirewallGroupDetail:
      type: object
      properties:
        firewall_group:
          $ref: '#/components/schemas/FirewallGroup'
...
    NewFirewallGroup:
      type: object
      properties:
        team_id:
          type: string
          format: uuid
        description:
          type: string
          maxLength: 255
          example: "Web servers firewall"
      required:
        - description
    UpdateFirewallGroup:
      type: object
      properties:
        team_id:
          type: string
          format: uuid
        description:
          type: string
          maxLength: 255
          example: "Updated firewall description"
      required:
        - description
...
    IPListSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Office IPs"
        description:
          type: string
          nullable: true
          example: "Office network addresses"
        entry_count:
          type: integer
          example: 5
        used_by_count:
          type: integer
          description: Number of firewall rules referencing this list
          example: 2
        created_at:
          type: string
          format: date-time
    IPLists:
      type: object
      properties:
        ip_lists:
          type: array
          items:
            $ref: '#/components/schemas/IPListSummary'
        ip_list_limit:
          type: integer
          description: Maximum number of IP lists allowed for this team
          example: 5
        ip_list_entry_limit:
          type: integer
          description: Maximum entries per IP list for this team
          example: 10
...
    IPListCreateResponse:
      type: object
      properties:
        ip_list:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            description:
              type: string
              nullable: true
    NewIPList:
      type: object
      properties:
        team_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 64
          example: "Office IPs"
        description:
          type: string
          maxLength: 255
          example: "Office network addresses"
      required:
        - name
...
  securitySchemes:
    api_key:
      type: http
      scheme: bearer